requirements.txt
Django==4.0.4
djangorestframework==3.13.1
stripe==3.0.0
psycopg2-binary